HW08 - updated scanner.sh to include testscript in the while loop to simulate files being added

+++ hw/hw08/testscript.sh
@@ -1,9 +1,19 @@
 #!/bin/sh
 
 toscan_dir="/escnfs/home/isetia/repos/scandata/toscan"
-test_dir="test"
+test_dir="./test"
 
-for archive in "$test_dir"/*; do
-	cp "$archive" "$toscan_dir"
-	sleep 2
+if [ ! -d "$test_dir" ]; then
+	echo "Direcotry $test_dir does not exist."
+	exit 1
+fi
+
+for archive in $test_dir/*; do
+	if [ -e "$archive" ]; then
+		cp "$archive" "$toscan_dir"
+		sleep 2
+	else 
+		echo "No files found in $test_dir"
+		break
+	fi
 done
